complted full rest-api with the database

+++ src/main/java/com/example/practice_2/crud/operations/Service/StudentService.java
@@ -44,7 +44,12 @@ public class StudentService {
 
     }
 
-    public Student deletestudentByid(int rno) {
-        return repo.deleteById(rno);
+    public String deletestudentByid(int rno) {
+           if(repo.existsById(rno)) {
+               repo.deleteById(rno);
+               return "Sucessfully deleted";
+           }
+
+           return "student data not found";
     }
 }
